Test DEV DAL

+++ ChessTournament_COMMON/Models/User.cs
@@ -13,10 +13,10 @@ namespace ChessTournament_COMMON.Models
             public string Usr_Pseudo { get; set; }
             public string Usr_Email { get; set; }
             public DateTime Usr_Birthdate{ get; set; }
-            public int Usr_Genre { get; set; }
+            public short Usr_Genre { get; set; }
             public string Usr_Firstname { get; set; }
             public string Usr_Lastname { get; set; }
-            public int Usr_Elo { get; set; }
+            public short Usr_Elo { get; set; }
             public bool Usr_Role { get; set; }
             public bool Usr_InActive { get; set; }
